Finished - for now ^_^
Finished the last two project pages, made sure each page has working links. The final project is difficult to implement into a webpage. The input required from the user is not intuitive and is easy to mess up. Its much easier to link the working environment at freeCodeCamp and allow the user to copy the solution to be used on fCC's website.